Refactor db package
